+++ plugin-cardea/src/main/java/ca/on/oicr/gsi/shesmu/cardea/CaseSummaryFormatDefinition.java
@@ -1,0 +1,12 @@
+package ca.on.oicr.gsi.shesmu.cardea;
+
+import ca.on.oicr.gsi.shesmu.plugin.input.InputFormat;
+import org.kohsuke.MetaInfServices;
+
+@MetaInfServices
+public class CaseSummaryFormatDefinition extends InputFormat {
+
+  public CaseSummaryFormatDefinition() {
+    super("case_summary", CaseSummaryValue.class);
+  }
+}